refactor(server): inline advisorMaxUses into openai_chat.go
Remove single-file mcp_integration.go and move advisorMaxUses()
into openai_chat.go where it belongs. This file was a merge
regression artifact containing only one helper function.
