+++ internal/server/openai_chat.go
@@ -757,3 +757,21 @@ func (s *Server) convertMessagesToResponseInputItems(messages []openai.ChatCompl
 func isValidRuleScenario(scenario typ.RuleScenario) bool {
 	return typ.CanUseScenarioInPath(scenario)
 }
+
+// advisorMaxUses returns the MaxUsesPerRequest from the advisor source config.
+// Returns 0 if no advisor is configured or the value is not positive.
+func (s *Server) advisorMaxUses() int {
+	if s.mcpRuntime == nil {
+		return 0
+	}
+	cfg := s.mcpRuntime.GetConfig()
+	if cfg == nil {
+		return 0
+	}
+	for _, source := range cfg.Sources {
+		if source.Advisor != nil && source.Advisor.MaxUsesPerRequest > 0 {
+			return source.Advisor.MaxUsesPerRequest
+		}
+	}
+	return 0
+}
